Handle also singlepart messsages
When a singlepart message is passed to the delivery method
an undefined method error is raised:

  undefined method `body' for nil:NilClass>

This is because the message class expects the content to be a mime-part.
In case of a singlepart message the content is in the message body.

This will use the body on singlepart messages and the text_part for
multipart messages.
It also extends the tests to cover the changed code.